Extract pytests from 'stream' test module into 'filehashstore' test module and delete redundant 'stream' test module
